fix(connectors): shared streaming size-cap reader for ado file hydration (promote from s3)
